functions/detect_labels/requirements.txt
google-cloud-firestore==0.30.0
google-cloud-vision==0.35.1